\name{psyc6130}
\alias{psyc6130}
\docType{package}
\title{This is a short package that includes all the function we will be using in this course.}

\description{
Contains functions and data sets.
}

\details{
  \tabular{ll}{
  Package: \tab psyc6130\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2010-11-09\cr
  License: \tab Freeware\cr
  LazyLoad: \tab yes\cr
  }
}

\author{Rob Cribbie and Phil Chalmers}
\keyword{ package }
