\docType{data}
\name{data_files}
\alias{data_files}
\title{Data sets used in the course}
\arguments{
  \item{contrasts}{Data-set for linear contrasts.}

  \item{nhl_data}{NHL data from 2010.}

  \item{twowaythreeway}{data for 2- and 3-way ANOVA.}

  \item{nonnorm_hetvar}{Non-normal data with heterogenous
  variance.}

  \item{one_way}{One way ANOVA data-set.}
}
\value{
  All data-sets are \code{data.frame} objects.
}
\description{
  A list of data files used throughout the course.
}
\author{
  Rob Cribbie and Phil Chalmers
}
\keyword{data}
\keyword{datasets}

