\docType{package}
\name{psyc6130}
\alias{psyc6130}
\alias{psyc6130-package}
\title{Package that includes all the function we will be using in this course.}
\description{
  Package that includes all the function we will be using
  in this course.
}
\details{
  Contains functions and data sets. \tabular{ll}{ Package:
  \tab psyc6130\cr Type: \tab Package\cr Version: \tab
  1.0\cr Date: \tab 2010-11-09\cr License: \tab Freeware\cr
  LazyLoad: \tab yes\cr }
}
\author{
  Rob Cribbie and Phil Chalmers
}
\keyword{package}

